LEARNING Remove about page and update blog and index pages
This commit deletes the `about.astro` page as it is no longer needed. It updates the `blog.astro` page with a new layout, navigation links, and content focused on learning Astro, including a list of posts. The `index.astro` page is revised to reflect personal information about Juan Pablo Flores, with custom styling and a dynamic title. These changes align the site with a more personalized and structured approach during the Astro learning journey.